Enrich data fetch subscriptions with additional JSON schema fields
